deleteFiles update

+++ R/deleteFiles.R
@@ -1,7 +1,10 @@
 setMethod(f = "deleteFiles",
-          signature = signature(x = "SimulationFiles"),
-          definition = function(x, trianglesInput = FALSE, maketOutput = FALSE) {
+          signature = signature(x = "SimulationFiles", deleteSimulationFiles = "boolean"),
+          definition = function(x, deleteSimulationFiles = FALSE,
+                                trianglesInput = FALSE, maketOutput = FALSE) {
               library(tools)
+              #here i am - sort out inputs for deleteFiles
+
               validObject(x, complete = TRUE)
               simDirs <- baseDir(x)
               for (i in 1:length(simDirs)) {
@@ -13,11 +16,14 @@ setMethod(f = "deleteFiles",
                   }
               }
 
-              fileNamesToDelete <- files(x)$fileName
-              if (product(x) == "images") {
-                  fileNamesToDelete <- paste0(tools::file_path_sans_ext(fileNamesToDelete), c(".grf", ".mpr", ".gr#"))
+              if (deleteSimulationFiles) {
+                  fileNamesToDelete <- files(x)$fileName
+                  if (product(x) == "images") {
+                      fileNamesToDelete <- paste0(tools::file_path_sans_ext(fileNamesToDelete),
+                                                  c(".grf", ".mpr", ".gr#"))
+                  }
+                  unlink(fileNamesToDelete, recursive = FALSE)
               }
-              unlink(fileNamesToDelete, recursive = FALSE)
 
               invisible(return(NULL))
           }
